+++ src/main/java/org/fuin/utils4j/filter/TokenPropertyFilter.java
@@ -17,6 +17,8 @@
  */
 package org.fuin.utils4j.filter;
 
+import org.jspecify.annotations.Nullable;
+
 /**
  * Defines a filter on a String property.
  */
@@ -45,7 +47,7 @@ public class TokenPropertyFilter extends PropertyFilter {
     }
 
     @Override
-    public final boolean complies(final Object obj) {
+    public final boolean complies(@Nullable final Object obj) {
         final String value = (String) this.getProperty(obj, this.getPropertyName());
         return filter.complies(value);
     }