+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Release Notes
 
-## 0.15.1
--  Added new [Utils4J](src/main/java/org/fuin/utils4j/Utils4J.java) method to encode a UUID to a short string and back 
+## 0.16.0
+- Added new [Utils4J](src/main/java/org/fuin/utils4j/Utils4J.java) method to encode a UUID to a short string and back
+- Added [JSpecify](https://jspecify.dev/) and [NullAway](https://github.com/uber/nullaway)
 
 ## 0.15.0
 - Added new [JandexUtils](src/main/java/org/fuin/utils4j/jandex/JandexUtils.java) methods